+++ lib/hooks/decimate.c
@@ -57,7 +57,7 @@ static int decimate_parse(struct hook *h, json_t *cfg)
 	return 0;
 }
 
-static int decimate_read(struct hook *h, struct sample *smps[], size_t *cnt)
+static int decimate_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 {
 	struct decimate *p = h->_vd;
