+++ lib/hooks/print.c
@@ -109,7 +109,7 @@ static int print_parse(struct hook *h, json_t *cfg)
 	return 0;
 }
 
-static int print_read(struct hook *h, struct sample *smps[], size_t *cnt)
+static int print_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 {
 	struct print *p = h->_vd;
